Refactor ProgCommandFactory and CommandFactory for LogCommand
