+++ src/main/java/command/programme/LogCommand.java
@@ -1,6 +1,7 @@
 package command.programme;
 import command.Command;
 import command.CommandResult;
+import dailyrecord.DailyRecord;
 import programme.ProgrammeList;
 import programme.Day;
 import history.History;
@@ -12,7 +13,7 @@ import java.util.logging.Logger;
 
 import static common.Utils.NULL_INTEGER;
 
-public class LogCommand extends Command {
+public class LogCommand extends ProgrammeCommand {
     public static final String COMMAND_WORD = "log";
     private final Logger logger = Logger.getLogger(this.getClass().getName());
 
@@ -52,7 +53,11 @@ public class LogCommand extends Command {
 
         assert completed != null : "Completed Day must not be null";
 
-        history.logDay(completed, date);
+        DailyRecord dailyRecord = history.getRecordByDate(date);
+
+        assert dailyRecord != null : "DailyRecord must not be null";
+
+        dailyRecord.logDay(completed);
 
         String result =  String.format("Congrats! You've successfully completed:%n%s",completed);
